+++ tests/testthat/test-get_datagrid.R
@@ -177,9 +177,15 @@ test_that("get_datagrid - data", {
     X = c("A", "A", "B"),
     Y = c(1, 5, 2),
     stringsAsFactors = FALSE
+  ), by = "Y", factors = "mode", length = 5)), 3L)
+
+  expect_identical(nrow(get_datagrid(data.frame(
+    X = c("A", "A", "B"),
+    Y = c(1.2, 5.5, 2),
+    stringsAsFactors = FALSE
   ), by = "Y", factors = "mode", length = 5)), 5L)
 
-  expect_identical(nrow(get_datagrid(iris, by = c("Sepal.Length = 3", "Species"))), 3L)
+    expect_identical(nrow(get_datagrid(iris, by = c("Sepal.Length = 3", "Species"))), 3L)
   expect_identical(nrow(get_datagrid(iris, by = c("Sepal.Length = c(3, 1)", "Species = 'setosa'"))), 2L)
 
   x1 <- get_datagrid(iris, by = c("Species", "Sepal.Length"), length = 30, preserve_range = TRUE)